+++ apiparser.h
@@ -12,11 +12,30 @@ struct ApiElementRecord {
     QString summary;
 };
 
+/**
+ * @brief Класс предназначен для разбора перечня элементов API по JSON-файлу
+ */
 class ApiParser
 {
 public:
-    ApiParser(const QByteArray &&text, size_t api_id);
+    /**
+     * @brief Конструирует класс ApiParser
+     * @param text JSON-файл, содержащий описание функций сервера
+     * @param api_id Номер API из базы данных
+     * Конструктор создаёт класс ApiParser на основе переданных ему входных данных.
+     * JSON-файл потенциально может быть объёмным, поэтому файл text перемещается
+     * во внутреннюю память объекта.
+     *
+     * Параметр api_id нужен для правильного заполнения поля \ref ApiElementRecord::api_id
+     * и последующего сохранения всех записей в базу данных.
+     */
+    ApiParser(const QByteArray &&text, qlonglong api_id);
 
+    /**
+     * @brief Функция выполняет разбор JSON-файла, переданного в конструкторе,
+     * извлекает из него все функции и возвращает их в виде списка структур
+     * \ref ApiElementRecord.
+     */
     QList<ApiElementRecord> parse();
 
 private:
